Review fixes: split nested types per file policy, Swift Testing, skill wording
Policy P2s from local autoreview: CodingAgent and LauncherError move to their
own extension files; the test suite converts from XCTest to Swift Testing.
The bundled skill mirrors the cmux-skills#9 CodeRabbit fixes: transport is
provider/image-dependent, vm open vs surface open targets, --json only where
documented.
